JPF:  * Adding support to 'SITE G[AL|MONTH|WK|DAY][UP|DN]'  * Adding 'SITE TRAFFIC' (reports general status of downloads/uploads).  * Fixed a leaky file descriptor in CommandInterface

+++ src/plugins/org.drftpd.commands.transferstatistics/src/org/drftpd/commands/transferstatistics/TransferStatistics.java
@@ -244,6 +244,7 @@ public class TransferStatistics extends CommandInterface  {
 
         return response;
     }
+    
     public CommandResponse doSITE_ALUP(CommandRequest request) {
     	return execute(request, "alup");
     }
@@ -379,7 +380,7 @@ public class TransferStatistics extends CommandInterface  {
     }
 
     public static String getUpRate(User user, int period) {
-        double s = user.getUploadedTimeForPeriod(period) / (double) 1000.0;
+        double s = user.getUploadedTimeForPeriod(period) / 1000;
 
         if (s <= 0) {
             return "- k/s";
@@ -391,7 +392,7 @@ public class TransferStatistics extends CommandInterface  {
     }
 
     public static String getDownRate(User user, int period) {
-        double s = user.getDownloadedTimeForPeriod(period) / (double) 1000.0;
+        double s = user.getDownloadedTimeForPeriod(period) / 1000;
 
         if (s <= 0) {
             return "- k/s";
